+++ tests/tinytest.R
@@ -1,0 +1,6 @@
+
+if ( requireNamespace("tinytest", quietly=TRUE) ){
+  library(tinytest)
+  library(anticlust)
+  tinytest::test_package("anticlust")
+}